+++ ChangeLog.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+
+
+- NEW : Add search field for custom ref in customer documents *14/09/2022* - 1.2.0
+- FIX : copy reference from source object to new object *26/09/2022* by [splohmer](https://github.com/splohmer)  
+  In case I create a custom reference in an offer (with or without the check to create it in the product) and then I transfer the offer to an order, the custom reference is lost.  
+  Same happens from order to invoice.  
+  Same for supplier and customer documents *
+
+
 ## Version 1.1
 
 - FIX : Onglet liste sql extrafields *12/06/2023* - 1.1.4